+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # v1.4.3
-## mm/dd/2019
+## 08/26/2022
 
 1. [](#new)
     * Pass phpstan level 1 tests